+++ mr/src/main/java/org/opensearch/hadoop/handler/impl/opensearch/BaseEventConverter.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.opensearch.hadoop.handler.impl.elasticsearch;
+package org.opensearch.hadoop.handler.impl.opensearch;
 
 import java.io.IOException;
 import java.util.Calendar;
