+++ qa/build.gradle
@@ -39,7 +39,7 @@ subprojects {
                 // For some reason the root dirs all point to the buildSrc folder. The local Repo will be one above that.
                 flatDir { dirs new File(project.rootProject.rootDir, "localRepo") }
             } else {
-                maven { url "https://oss.sonatype.org/content/repositories/snapshots" }
+                maven { url "https://aws.oss.sonatype.org/content/repositories/snapshots" }
             }
         }
         dependencies {
